% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{diffusionMap}
\alias{diffusionMap}
\title{Diffusion model}
\usage{
diffusionMap(receptors, TFs, M, network, nCores = 2, nTicks = 2000)
}
\arguments{
\item{receptors}{vector of string with ids.}

\item{TFs}{vector of string with gene ids of Transcription Factors}

\item{network}{a dataframe of two columns "from" and "to" with strings representing gene IDs}

\item{nCores}{number of cores for parallel processing}

\item{nTicks}{maximum number of time steps. Defaults to 2000.}
}
\value{
list of two components: Enode which is end node and Signal i.e. final state of network
}
\description{
This function calculates single-sample network connectivity by using aggregate accumulation of signal per node as a function of time.
}
\examples{
data(diffusionData)
diffusionMap <- function(receptors, TFs, M, network, nCores=2, nTicks=2000)
{ ... }
signalNet
}
\keyword{Diffusion}
\keyword{Model}
