% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{getSubnetwork}
\alias{getSubnetwork}
\title{Network Graph}
\usage{
getSubnetwork(network, receptors, TFs)
}
\arguments{
\item{network}{a dataframe of two columns "from" and "to" with strings representing gene IDs}

\item{receptors}{receptors in the network node}

\item{TFs}{transcription factors in the network node}
}
\value{
list of two components: network subgraph from receptor i.e. start node to TF i.e. end node
}
\description{
This function generates network graph from receptor to TF by using shortest path with neighboring network nodes.
}
\examples{
# g <- graph_from_data_frame(network, directed = F, vertices = NULL)
#{ ... }
}
\keyword{getSubnetwork}
