% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{initializeSignallingNetwork}
\alias{initializeSignallingNetwork}
\title{receptor2tfDiffusion}
\usage{
initializeSignallingNetwork(network, nodeW)
}
\arguments{
\item{network}{a dataframe of two columns "from" and "to" with strings representing gene IDs}

\item{nodeW}{weight on the network node}
}
\value{

}
\description{
To calculate Network diffusion, there are mainly two steps: creating Signalling Network (SigNet) and
use SigNet output in calculate single-sample network connectivity using diffusion model.
}
\examples{
#network <- initializeSignallingNetwork(network = network, nodeW = nodeW)
#{ ... }
}
