% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{signalOnNetwork}
\alias{signalOnNetwork}
\title{Diffusion model}
\usage{
signalOnNetwork(
  network,
  nodeW,
  outputNode = "FLI1",
  inputNode = "TNFRSF1B",
  inputSignal = 0.99,
  n = 2000
)
}
\arguments{
\item{network}{a dataframe of two columns "from" and "to" with strings representing gene IDs}

\item{outputNode}{output node}

\item{inputNode}{input node containing receptors in the signalling network}

\item{inputSignal}{minimum amount of signal placed as input for signal propagation}

\item{n}{number of iteration}
}
\value{

}
\description{
To calculate Network diffusion, there are mainly two steps: creating Signalling Network (SigNet) and
use SigNet output in calculate single-sample network connectivity using diffusion model.
}
\details{
Signalling Network

Signalling Network quantify network connectivity by calculating signal strength between receptors and TFs.
}
\examples{
toydata <- data(diffusionData)
signalNet<- signalOnNetwork(diffusionData$network, nodeW=M[,1], outputNode = 'FLI1',inputNode = 'TNFRSF1B', inputSignal = 0.99,n = 2000)
{ ... }
}
\keyword{Network}
\keyword{Signalling}
