% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion.R
\name{signalOnNetwork}
\alias{signalOnNetwork}
\title{Quantify network connectivity by calculating signal strength between receptors and TFs.}
\usage{
signalOnNetwork(network, outputNode, inputNode, inputSignal = 0.99, n = 2000)
}
\arguments{
\item{network}{output from initilizeNetwork list with 'network' and 'signal' components}

\item{outputNode}{output node}

\item{inputNode}{input node containing receptors in the signalling network}

\item{inputSignal}{minimum amount of signal placed as input for signal propagation}

\item{n}{number of iteration}

\item{nodeW}{weight on the network node}
}
\value{

}
\description{
Quantify network connectivity by calculating signal strength between receptors and TFs.
}
\examples{
#data(diffusionData)
#network <- initializeSignallingNetwork(network = network, nodeW = nodeW)
#signalNet<- signalOnNetwork(diffusionData$network, nodeW=M[,1], outputNode = 'FLI1',inputNode = 'TNFRSF1B', inputSignal = 0.99,n = 2000)
#{ ... }
}
\keyword{Network}
\keyword{Signalling}
